workspace/requirements.txt
psycopg2-binary
bokeh
chart_studio
numpy
scipy
python-dotenv